LayoutTests/js/dfg-get-by-val-getter-cse-expected.txt

This tests that common subexpression elimination knows to be conservative when getters are involved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS doSomeGets(array) is 77
PASS counter is 3
PASS array[1] is 3
PASS doSomeGets(array) is 79
PASS counter is 4
PASS array[1] is 4
PASS doSomeGets(array) is 81
PASS counter is 5
PASS array[1] is 5
PASS doSomeGets(array) is 83
PASS counter is 6
PASS array[1] is 6
PASS doSomeGets(array) is 85
PASS counter is 7
PASS array[1] is 7
PASS doSomeGets(array) is 87
PASS counter is 8
PASS array[1] is 8
PASS doSomeGets(array) is 89
PASS counter is 9
PASS array[1] is 9
PASS doSomeGets(array) is 91
PASS counter is 10
PASS array[1] is 10
PASS doSomeGets(array) is 93
PASS counter is 11
PASS array[1] is 11
PASS doSomeGets(array) is 95
PASS counter is 12
PASS array[1] is 12
PASS doSomeGets(array) is 97
PASS counter is 13
PASS array[1] is 13
PASS doSomeGets(array) is 99
PASS counter is 14
PASS array[1] is 14
PASS doSomeGets(array) is 101
PASS counter is 15
PASS array[1] is 15
PASS doSomeGets(array) is 103
PASS counter is 16
PASS array[1] is 16
PASS doSomeGets(array) is 105
PASS counter is 17
PASS array[1] is 17
PASS doSomeGets(array) is 107
PASS counter is 18
PASS array[1] is 18
PASS doSomeGets(array) is 109
PASS counter is 19
PASS array[1] is 19
PASS doSomeGets(array) is 111
PASS counter is 20
PASS array[1] is 20
PASS doSomeGets(array) is 113
PASS counter is 21
PASS array[1] is 21
PASS doSomeGets(array) is 115
PASS counter is 22
PASS array[1] is 22
PASS doSomeGets(array) is 117
PASS counter is 23
PASS array[1] is 23
PASS doSomeGets(array) is 119
PASS counter is 24
PASS array[1] is 24
PASS doSomeGets(array) is 121
PASS counter is 25
PASS array[1] is 25
PASS doSomeGets(array) is 123
PASS counter is 26
PASS array[1] is 26
PASS doSomeGets(array) is 125
PASS counter is 27
PASS array[1] is 27
PASS doSomeGets(array) is 127
PASS counter is 28
PASS array[1] is 28
PASS doSomeGets(array) is 129
PASS counter is 29
PASS array[1] is 29
PASS doSomeGets(array) is 131
PASS counter is 30
PASS array[1] is 30
PASS doSomeGets(array) is 133
PASS counter is 31
PASS array[1] is 31
PASS doSomeGets(array) is 135
PASS counter is 32
PASS array[1] is 32
PASS doSomeGets(array) is 137
PASS counter is 33
PASS array[1] is 33
PASS doSomeGets(array) is 139
PASS counter is 34
PASS array[1] is 34
PASS doSomeGets(array) is 141
PASS counter is 35
PASS array[1] is 35
PASS doSomeGets(array) is 143
PASS counter is 36
PASS array[1] is 36
PASS doSomeGets(array) is 145
PASS counter is 37
PASS array[1] is 37
PASS doSomeGets(array) is 147
PASS counter is 38
PASS array[1] is 38
PASS doSomeGets(array) is 149
PASS counter is 39
PASS array[1] is 39
PASS doSomeGets(array) is 151
PASS counter is 40
PASS array[1] is 40
PASS doSomeGets(array) is 153
PASS counter is 41
PASS array[1] is 41
PASS doSomeGets(array) is 155
PASS counter is 42
PASS array[1] is 42
PASS doSomeGets(array) is 157
PASS counter is 43
PASS array[1] is 43
PASS doSomeGets(array) is 159
PASS counter is 44
PASS array[1] is 44
PASS doSomeGets(array) is 161
PASS counter is 45
PASS array[1] is 45
PASS doSomeGets(array) is 163
PASS counter is 46
PASS array[1] is 46
PASS doSomeGets(array) is 165
PASS counter is 47
PASS array[1] is 47
PASS doSomeGets(array) is 167
PASS counter is 48
PASS array[1] is 48
PASS doSomeGets(array) is 169
PASS counter is 49
PASS array[1] is 49
PASS doSomeGets(array) is 171
PASS counter is 50
PASS array[1] is 50
PASS doSomeGets(array) is 173
PASS counter is 51
PASS array[1] is 51
PASS doSomeGets(array) is 175
PASS counter is 52
PASS array[1] is 52
PASS doSomeGets(array) is 177
PASS counter is 53
PASS array[1] is 53
PASS doSomeGets(array) is 179
PASS counter is 54
PASS array[1] is 54
PASS doSomeGets(array) is 181
PASS counter is 55
PASS array[1] is 55
PASS doSomeGets(array) is 183
PASS counter is 56
PASS array[1] is 56
PASS doSomeGets(array) is 185
PASS counter is 57
PASS array[1] is 57
PASS doSomeGets(array) is 187
PASS counter is 58
PASS array[1] is 58
PASS doSomeGets(array) is 189
PASS counter is 59
PASS array[1] is 59
PASS doSomeGets(array) is 191
PASS counter is 60
PASS array[1] is 60
PASS doSomeGets(array) is 193
PASS counter is 61
PASS array[1] is 61
PASS doSomeGets(array) is 195
PASS counter is 62
PASS array[1] is 62
PASS doSomeGets(array) is 197
PASS counter is 63
PASS array[1] is 63
PASS doSomeGets(array) is 199
PASS counter is 64
PASS array[1] is 64
PASS doSomeGets(array) is 201
PASS counter is 65
PASS array[1] is 65
PASS doSomeGets(array) is 203
PASS counter is 66
PASS array[1] is 66
PASS doSomeGets(array) is 205
PASS counter is 67
PASS array[1] is 67
PASS doSomeGets(array) is 207
PASS counter is 68
PASS array[1] is 68
PASS doSomeGets(array) is 209
PASS counter is 69
PASS array[1] is 69
PASS doSomeGets(array) is 211
PASS counter is 70
PASS array[1] is 70
PASS doSomeGets(array) is 213
PASS counter is 71
PASS array[1] is 71
PASS doSomeGets(array) is 215
PASS counter is 72
PASS array[1] is 72
PASS doSomeGets(array) is 217
PASS counter is 73
PASS array[1] is 73
PASS doSomeGets(array) is 219
PASS counter is 74
PASS array[1] is 74
PASS doSomeGets(array) is 221
PASS counter is 75
PASS array[1] is 75
PASS doSomeGets(array) is 223
PASS counter is 76
PASS array[1] is 76
PASS doSomeGets(array) is 225
PASS counter is 77
PASS array[1] is 77
PASS doSomeGets(array) is 227
PASS counter is 78
PASS array[1] is 78
PASS doSomeGets(array) is 229
PASS counter is 79
PASS array[1] is 79
PASS doSomeGets(array) is 231
PASS counter is 80
PASS array[1] is 80
PASS doSomeGets(array) is 233
PASS counter is 81
PASS array[1] is 81
PASS doSomeGets(array) is 235
PASS counter is 82
PASS array[1] is 82
PASS doSomeGets(array) is 237
PASS counter is 83
PASS array[1] is 83
PASS doSomeGets(array) is 239
PASS counter is 84
PASS array[1] is 84
PASS doSomeGets(array) is 241
PASS counter is 85
PASS array[1] is 85
PASS doSomeGets(array) is 243
PASS counter is 86
PASS array[1] is 86
PASS doSomeGets(array) is 245
PASS counter is 87
PASS array[1] is 87
PASS doSomeGets(array) is 247
PASS counter is 88
PASS array[1] is 88
PASS doSomeGets(array) is 249
PASS counter is 89
PASS array[1] is 89
PASS doSomeGets(array) is 251
PASS counter is 90
PASS array[1] is 90
PASS doSomeGets(array) is 253
PASS counter is 91
PASS array[1] is 91
PASS doSomeGets(array) is 255
PASS counter is 92
PASS array[1] is 92
PASS doSomeGets(array) is 257
PASS counter is 93
PASS array[1] is 93
PASS doSomeGets(array) is 259
PASS counter is 94
PASS array[1] is 94
PASS doSomeGets(array) is 261
PASS counter is 95
PASS array[1] is 95
PASS doSomeGets(array) is 263
PASS counter is 96
PASS array[1] is 96
PASS doSomeGets(array) is 265
PASS counter is 97
PASS array[1] is 97
PASS doSomeGets(array) is 267
PASS counter is 98
PASS array[1] is 98
PASS doSomeGets(array) is 269
PASS counter is 99
PASS array[1] is 99
PASS doSomeGets(array) is 271
PASS counter is 100
PASS array[1] is 100
PASS doSomeGets(array) is 273
PASS counter is 101
PASS array[1] is 101
PASS doSomeGets(array) is 275
PASS counter is 102
PASS array[1] is 102
PASS doSomeGets(array) is 277
PASS counter is 103
PASS array[1] is 103
PASS doSomeGets(array) is 279
PASS counter is 104
PASS array[1] is 104
PASS doSomeGets(array) is 281
PASS counter is 105
PASS array[1] is 105
PASS doSomeGets(array) is 283
PASS counter is 106
PASS array[1] is 106
PASS doSomeGets(array) is 285
PASS counter is 107
PASS array[1] is 107
PASS doSomeGets(array) is 287
PASS counter is 108
PASS array[1] is 108
PASS doSomeGets(array) is 289
PASS counter is 109
PASS array[1] is 109
PASS doSomeGets(array) is 291
PASS counter is 110
PASS array[1] is 110
PASS doSomeGets(array) is 293
PASS counter is 111
PASS array[1] is 111
PASS doSomeGets(array) is 295
PASS counter is 112
PASS array[1] is 112
PASS doSomeGets(array) is 297
PASS counter is 113
PASS array[1] is 113
PASS doSomeGets(array) is 299
PASS counter is 114
PASS array[1] is 114
PASS doSomeGets(array) is 301
PASS counter is 115
PASS array[1] is 115
PASS doSomeGets(array) is 303
PASS counter is 116
PASS array[1] is 116
PASS doSomeGets(array) is 305
PASS counter is 117
PASS array[1] is 117
PASS doSomeGets(array) is 307
PASS counter is 118
PASS array[1] is 118
PASS doSomeGets(array) is 309
PASS counter is 119
PASS array[1] is 119
PASS doSomeGets(array) is 311
PASS counter is 120
PASS array[1] is 120
PASS doSomeGets(array) is 313
PASS counter is 121
PASS array[1] is 121
PASS doSomeGets(array) is 315
PASS counter is 122
PASS array[1] is 122
PASS doSomeGets(array) is 317
PASS counter is 123
PASS array[1] is 123
PASS doSomeGets(array) is 319
PASS counter is 124
PASS array[1] is 124
PASS doSomeGets(array) is 321
PASS counter is 125
PASS array[1] is 125
PASS doSomeGets(array) is 323
PASS counter is 126
PASS array[1] is 126
PASS doSomeGets(array) is 325
PASS counter is 127
PASS array[1] is 127
PASS doSomeGets(array) is 327
PASS counter is 128
PASS array[1] is 128
PASS doSomeGets(array) is 329
PASS counter is 129
PASS array[1] is 129
PASS doSomeGets(array) is 331
PASS counter is 130
PASS array[1] is 130
PASS doSomeGets(array) is 333
PASS counter is 131
PASS array[1] is 131
PASS doSomeGets(array) is 335
PASS counter is 132
PASS array[1] is 132
PASS doSomeGets(array) is 337
PASS counter is 133
PASS array[1] is 133
PASS doSomeGets(array) is 339
PASS counter is 134
PASS array[1] is 134
PASS doSomeGets(array) is 341
PASS counter is 135
PASS array[1] is 135
PASS doSomeGets(array) is 343
PASS counter is 136
PASS array[1] is 136
PASS doSomeGets(array) is 345
PASS counter is 137
PASS array[1] is 137
PASS doSomeGets(array) is 347
PASS counter is 138
PASS array[1] is 138
PASS doSomeGets(array) is 349
PASS counter is 139
PASS array[1] is 139
PASS doSomeGets(array) is 351
PASS counter is 140
PASS array[1] is 140
PASS doSomeGets(array) is 353
PASS counter is 141
PASS array[1] is 141
PASS doSomeGets(array) is 355
PASS counter is 142
PASS array[1] is 142
PASS doSomeGets(array) is 357
PASS counter is 143
PASS array[1] is 143
PASS doSomeGets(array) is 359
PASS counter is 144
PASS array[1] is 144
PASS doSomeGets(array) is 361
PASS counter is 145
PASS array[1] is 145
PASS doSomeGets(array) is 363
PASS counter is 146
PASS array[1] is 146
PASS doSomeGets(array) is 365
PASS counter is 147
PASS array[1] is 147
PASS doSomeGets(array) is 367
PASS counter is 148
PASS array[1] is 148
PASS doSomeGets(array) is 369
PASS counter is 149
PASS array[1] is 149
PASS doSomeGets(array) is 371
PASS counter is 150
PASS array[1] is 150
PASS doSomeGets(array) is 373
PASS counter is 151
PASS array[1] is 151
PASS doSomeGets(array) is 375
PASS counter is 152
PASS array[1] is 152
PASS doSomeGets(array) is 377
PASS counter is 153
PASS array[1] is 153
PASS doSomeGets(array) is 379
PASS counter is 154
PASS array[1] is 154
PASS doSomeGets(array) is 381
PASS counter is 155
PASS array[1] is 155
PASS doSomeGets(array) is 383
PASS counter is 156
PASS array[1] is 156
PASS doSomeGets(array) is 385
PASS counter is 157
PASS array[1] is 157
PASS doSomeGets(array) is 387
PASS counter is 158
PASS array[1] is 158
PASS doSomeGets(array) is 389
PASS counter is 159
PASS array[1] is 159
PASS doSomeGets(array) is 391
PASS counter is 160
PASS array[1] is 160
PASS doSomeGets(array) is 393
PASS counter is 161
PASS array[1] is 161
PASS doSomeGets(array) is 395
PASS counter is 162
PASS array[1] is 162
PASS doSomeGets(array) is 397
PASS counter is 163
PASS array[1] is 163
PASS doSomeGets(array) is 399
PASS counter is 164
PASS array[1] is 164
PASS doSomeGets(array) is 401
PASS counter is 165
PASS array[1] is 165
PASS doSomeGets(array) is 403
PASS counter is 166
PASS array[1] is 166
PASS doSomeGets(array) is 405
PASS counter is 167
PASS array[1] is 167
PASS doSomeGets(array) is 407
PASS counter is 168
PASS array[1] is 168
PASS doSomeGets(array) is 409
PASS counter is 169
PASS array[1] is 169
PASS doSomeGets(array) is 411
PASS counter is 170
PASS array[1] is 170
PASS doSomeGets(array) is 413
PASS counter is 171
PASS array[1] is 171
PASS doSomeGets(array) is 415
PASS counter is 172
PASS array[1] is 172
PASS doSomeGets(array) is 417
PASS counter is 173
PASS array[1] is 173
PASS doSomeGets(array) is 419
PASS counter is 174
PASS array[1] is 174
PASS doSomeGets(array) is 421
PASS counter is 175
PASS array[1] is 175
PASS doSomeGets(array) is 423
PASS counter is 176
PASS array[1] is 176
PASS doSomeGets(array) is 425
PASS counter is 177
PASS array[1] is 177
PASS doSomeGets(array) is 427
PASS counter is 178
PASS array[1] is 178
PASS doSomeGets(array) is 429
PASS counter is 179
PASS array[1] is 179
PASS doSomeGets(array) is 431
PASS counter is 180
PASS array[1] is 180
PASS doSomeGets(array) is 433
PASS counter is 181
PASS array[1] is 181
PASS doSomeGets(array) is 435
PASS counter is 182
PASS array[1] is 182
PASS doSomeGets(array) is 437
PASS counter is 183
PASS array[1] is 183
PASS doSomeGets(array) is 439
PASS counter is 184
PASS array[1] is 184
PASS doSomeGets(array) is 441
PASS counter is 185
PASS array[1] is 185
PASS doSomeGets(array) is 443
PASS counter is 186
PASS array[1] is 186
PASS doSomeGets(array) is 445
PASS counter is 187
PASS array[1] is 187
PASS doSomeGets(array) is 447
PASS counter is 188
PASS array[1] is 188
PASS doSomeGets(array) is 449
PASS counter is 189
PASS array[1] is 189
PASS doSomeGets(array) is 451
PASS counter is 190
PASS array[1] is 190
PASS doSomeGets(array) is 453
PASS counter is 191
PASS array[1] is 191
PASS doSomeGets(array) is 455
PASS counter is 192
PASS array[1] is 192
PASS doSomeGets(array) is 457
PASS counter is 193
PASS array[1] is 193
PASS doSomeGets(array) is 459
PASS counter is 194
PASS array[1] is 194
PASS doSomeGets(array) is 461
PASS counter is 195
PASS array[1] is 195
PASS doSomeGets(array) is 463
PASS counter is 196
PASS array[1] is 196
PASS doSomeGets(array) is 465
PASS counter is 197
PASS array[1] is 197
PASS doSomeGets(array) is 467
PASS counter is 198
PASS array[1] is 198
PASS doSomeGets(array) is 469
PASS counter is 199
PASS array[1] is 199
PASS doSomeGets(array) is 471
PASS counter is 200
PASS array[1] is 200
PASS doSomeGets(array) is 473
PASS counter is 201
PASS array[1] is 201
PASS doSomeGets(array) is 475
PASS counter is 202
PASS array[1] is 202
PASS doSomeGets(array) is 477
PASS counter is 203
PASS array[1] is 203
PASS doSomeGets(array) is 479
PASS counter is 204
PASS array[1] is 204
PASS doSomeGets(array) is 481
PASS counter is 205
PASS array[1] is 205
PASS doSomeGets(array) is 483
PASS counter is 206
PASS array[1] is 206
PASS doSomeGets(array) is 485
PASS counter is 207
PASS array[1] is 207
PASS doSomeGets(array) is 487
PASS counter is 208
PASS array[1] is 208
PASS doSomeGets(array) is 489
PASS counter is 209
PASS array[1] is 209
PASS doSomeGets(array) is 491
PASS counter is 210
PASS array[1] is 210
PASS doSomeGets(array) is 493
PASS counter is 211
PASS array[1] is 211
PASS doSomeGets(array) is 495
PASS counter is 212
PASS array[1] is 212
PASS doSomeGets(array) is 497
PASS counter is 213
PASS array[1] is 213
PASS doSomeGets(array) is 499
PASS counter is 214
PASS array[1] is 214
PASS doSomeGets(array) is 501
PASS counter is 215
PASS array[1] is 215
PASS doSomeGets(array) is 503
PASS counter is 216
PASS array[1] is 216
PASS doSomeGets(array) is 505
PASS counter is 217
PASS array[1] is 217
PASS doSomeGets(array) is 507
PASS counter is 218
PASS array[1] is 218
PASS doSomeGets(array) is 509
PASS counter is 219
PASS array[1] is 219
PASS doSomeGets(array) is 511
PASS counter is 220
PASS array[1] is 220
PASS doSomeGets(array) is 513
PASS counter is 221
PASS array[1] is 221
PASS doSomeGets(array) is 515
PASS counter is 222
PASS array[1] is 222
PASS doSomeGets(array) is 517
PASS counter is 223
PASS array[1] is 223
PASS doSomeGets(array) is 519
PASS counter is 224
PASS array[1] is 224
PASS doSomeGets(array) is 521
PASS counter is 225
PASS array[1] is 225
PASS doSomeGets(array) is 523
PASS counter is 226
PASS array[1] is 226
PASS doSomeGets(array) is 525
PASS counter is 227
PASS array[1] is 227
PASS doSomeGets(array) is 527
PASS counter is 228
PASS array[1] is 228
PASS doSomeGets(array) is 529
PASS counter is 229
PASS array[1] is 229
PASS doSomeGets(array) is 531
PASS counter is 230
PASS array[1] is 230
PASS doSomeGets(array) is 533
PASS counter is 231
PASS array[1] is 231
PASS doSomeGets(array) is 535
PASS counter is 232
PASS array[1] is 232
PASS doSomeGets(array) is 537
PASS counter is 233
PASS array[1] is 233
PASS doSomeGets(array) is 539
PASS counter is 234
PASS array[1] is 234
PASS doSomeGets(array) is 541
PASS counter is 235
PASS array[1] is 235
PASS doSomeGets(array) is 543
PASS counter is 236
PASS array[1] is 236
PASS doSomeGets(array) is 545
PASS counter is 237
PASS array[1] is 237
PASS doSomeGets(array) is 547
PASS counter is 238
PASS array[1] is 238
PASS doSomeGets(array) is 549
PASS counter is 239
PASS array[1] is 239
PASS doSomeGets(array) is 551
PASS counter is 240
PASS array[1] is 240
PASS doSomeGets(array) is 553
PASS counter is 241
PASS array[1] is 241
PASS doSomeGets(array) is 555
PASS counter is 242
PASS array[1] is 242
PASS doSomeGets(array) is 557
PASS counter is 243
PASS array[1] is 243
PASS doSomeGets(array) is 559
PASS counter is 244
PASS array[1] is 244
PASS doSomeGets(array) is 561
PASS counter is 245
PASS array[1] is 245
PASS doSomeGets(array) is 563
PASS counter is 246
PASS array[1] is 246
PASS doSomeGets(array) is 565
PASS counter is 247
PASS array[1] is 247
PASS doSomeGets(array) is 567
PASS counter is 248
PASS array[1] is 248
PASS doSomeGets(array) is 569
PASS counter is 249
PASS array[1] is 249
PASS doSomeGets(array) is 571
PASS counter is 250
PASS array[1] is 250
PASS doSomeGets(array) is 573
PASS counter is 251
PASS array[1] is 251
PASS doSomeGets(array) is 575
PASS counter is 252
PASS array[1] is 252
PASS doSomeGets(array) is 577
PASS counter is 253
PASS array[1] is 253
PASS doSomeGets(array) is 579
PASS counter is 254
PASS array[1] is 254
PASS doSomeGets(array) is 581
PASS counter is 255
PASS array[1] is 255
PASS doSomeGets(array) is 583
PASS counter is 256
PASS array[1] is 256
PASS doSomeGets(array) is 585
PASS counter is 257
PASS array[1] is 257
PASS doSomeGets(array) is 587
PASS counter is 258
PASS array[1] is 258
PASS doSomeGets(array) is 589
PASS counter is 259
PASS array[1] is 259
PASS doSomeGets(array) is 591
PASS counter is 260
PASS array[1] is 260
PASS doSomeGets(array) is 593
PASS counter is 261
PASS array[1] is 261
PASS doSomeGets(array) is 595
PASS counter is 262
PASS array[1] is 262
PASS doSomeGets(array) is 597
PASS counter is 263
PASS array[1] is 263
PASS doSomeGets(array) is 599
PASS counter is 264
PASS array[1] is 264
PASS doSomeGets(array) is 601
PASS counter is 265
PASS array[1] is 265
PASS doSomeGets(array) is 603
PASS counter is 266
PASS array[1] is 266
PASS doSomeGets(array) is 605
PASS counter is 267
PASS array[1] is 267
PASS doSomeGets(array) is 607
PASS counter is 268
PASS array[1] is 268
PASS doSomeGets(array) is 609
PASS counter is 269
PASS array[1] is 269
PASS doSomeGets(array) is 611
PASS counter is 270
PASS array[1] is 270
PASS doSomeGets(array) is 613
PASS counter is 271
PASS array[1] is 271
PASS doSomeGets(array) is 615
PASS counter is 272
PASS array[1] is 272
PASS doSomeGets(array) is 617
PASS counter is 273
PASS array[1] is 273
PASS doSomeGets(array) is 619
PASS counter is 274
PASS array[1] is 274
PASS doSomeGets(array) is 621
PASS counter is 275
PASS array[1] is 275
PASS doSomeGets(array) is 623
PASS counter is 276
PASS array[1] is 276
PASS doSomeGets(array) is 625
PASS counter is 277
PASS array[1] is 277
PASS doSomeGets(array) is 627
PASS counter is 278
PASS array[1] is 278
PASS doSomeGets(array) is 629
PASS counter is 279
PASS array[1] is 279
PASS doSomeGets(array) is 631
PASS counter is 280
PASS array[1] is 280
PASS doSomeGets(array) is 633
PASS counter is 281
PASS array[1] is 281
PASS doSomeGets(array) is 635
PASS counter is 282
PASS array[1] is 282
PASS doSomeGets(array) is 637
PASS counter is 283
PASS array[1] is 283
PASS doSomeGets(array) is 639
PASS counter is 284
PASS array[1] is 284
PASS doSomeGets(array) is 641
PASS counter is 285
PASS array[1] is 285
PASS doSomeGets(array) is 643
PASS counter is 286
PASS array[1] is 286
PASS doSomeGets(array) is 645
PASS counter is 287
PASS array[1] is 287
PASS doSomeGets(array) is 647
PASS counter is 288
PASS array[1] is 288
PASS doSomeGets(array) is 649
PASS counter is 289
PASS array[1] is 289
PASS doSomeGets(array) is 651
PASS counter is 290
PASS array[1] is 290
PASS doSomeGets(array) is 653
PASS counter is 291
PASS array[1] is 291
PASS doSomeGets(array) is 655
PASS counter is 292
PASS array[1] is 292
PASS doSomeGets(array) is 657
PASS counter is 293
PASS array[1] is 293
PASS doSomeGets(array) is 659
PASS counter is 294
PASS array[1] is 294
PASS doSomeGets(array) is 661
PASS counter is 295
PASS array[1] is 295
PASS doSomeGets(array) is 663
PASS counter is 296
PASS array[1] is 296
PASS doSomeGets(array) is 665
PASS counter is 297
PASS array[1] is 297
PASS doSomeGets(array) is 667
PASS counter is 298
PASS array[1] is 298
PASS doSomeGets(array) is 669
PASS counter is 299
PASS array[1] is 299
PASS doSomeGets(array) is 671
PASS counter is 300
PASS array[1] is 300
PASS doSomeGets(array) is 673
PASS counter is 301
PASS array[1] is 301
PASS doSomeGets(array) is 675
PASS counter is 302
PASS array[1] is 302
PASS doSomeGets(array) is 677
PASS counter is 303
PASS array[1] is 303
PASS doSomeGets(array) is 679
PASS counter is 304
PASS array[1] is 304
PASS doSomeGets(array) is 681
PASS counter is 305
PASS array[1] is 305
PASS doSomeGets(array) is 683
PASS counter is 306
PASS array[1] is 306
PASS doSomeGets(array) is 685
PASS counter is 307
PASS array[1] is 307
PASS doSomeGets(array) is 687
PASS counter is 308
PASS array[1] is 308
PASS doSomeGets(array) is 689
PASS counter is 309
PASS array[1] is 309
PASS doSomeGets(array) is 691
PASS counter is 310
PASS array[1] is 310
PASS doSomeGets(array) is 693
PASS counter is 311
PASS array[1] is 311
PASS doSomeGets(array) is 695
PASS counter is 312
PASS array[1] is 312
PASS doSomeGets(array) is 697
PASS counter is 313
PASS array[1] is 313
PASS doSomeGets(array) is 699
PASS counter is 314
PASS array[1] is 314
PASS doSomeGets(array) is 701
PASS counter is 315
PASS array[1] is 315
PASS doSomeGets(array) is 703
PASS counter is 316
PASS array[1] is 316
PASS doSomeGets(array) is 705
PASS counter is 317
PASS array[1] is 317
PASS doSomeGets(array) is 707
PASS counter is 318
PASS array[1] is 318
PASS doSomeGets(array) is 709
PASS counter is 319
PASS array[1] is 319
PASS doSomeGets(array) is 711
PASS counter is 320
PASS array[1] is 320
PASS doSomeGets(array) is 713
PASS counter is 321
PASS array[1] is 321
PASS doSomeGets(array) is 715
PASS counter is 322
PASS array[1] is 322
PASS doSomeGets(array) is 717
PASS counter is 323
PASS array[1] is 323
PASS doSomeGets(array) is 719
PASS counter is 324
PASS array[1] is 324
PASS doSomeGets(array) is 721
PASS counter is 325
PASS array[1] is 325
PASS doSomeGets(array) is 723
PASS counter is 326
PASS array[1] is 326
PASS doSomeGets(array) is 725
PASS counter is 327
PASS array[1] is 327
PASS doSomeGets(array) is 727
PASS counter is 328
PASS array[1] is 328
PASS doSomeGets(array) is 729
PASS counter is 329
PASS array[1] is 329
PASS doSomeGets(array) is 731
PASS counter is 330
PASS array[1] is 330
PASS doSomeGets(array) is 733
PASS counter is 331
PASS array[1] is 331
PASS doSomeGets(array) is 735
PASS counter is 332
PASS array[1] is 332
PASS doSomeGets(array) is 737
PASS counter is 333
PASS array[1] is 333
PASS doSomeGets(array) is 739
PASS counter is 334
PASS array[1] is 334
PASS doSomeGets(array) is 741
PASS counter is 335
PASS array[1] is 335
PASS doSomeGets(array) is 743
PASS counter is 336
PASS array[1] is 336
PASS doSomeGets(array) is 745
PASS counter is 337
PASS array[1] is 337
PASS doSomeGets(array) is 747
PASS counter is 338
PASS array[1] is 338
PASS doSomeGets(array) is 749
PASS counter is 339
PASS array[1] is 339
PASS doSomeGets(array) is 751
PASS counter is 340
PASS array[1] is 340
PASS doSomeGets(array) is 753
PASS counter is 341
PASS array[1] is 341
PASS doSomeGets(array) is 755
PASS counter is 342
PASS array[1] is 342
PASS doSomeGets(array) is 757
PASS counter is 343
PASS array[1] is 343
PASS doSomeGets(array) is 759
PASS counter is 344
PASS array[1] is 344
PASS doSomeGets(array) is 761
PASS counter is 345
PASS array[1] is 345
PASS doSomeGets(array) is 763
PASS counter is 346
PASS array[1] is 346
PASS doSomeGets(array) is 765
PASS counter is 347
PASS array[1] is 347
PASS doSomeGets(array) is 767
PASS counter is 348
PASS array[1] is 348
PASS doSomeGets(array) is 769
PASS counter is 349
PASS array[1] is 349
PASS doSomeGets(array) is 771
PASS counter is 350
PASS array[1] is 350
PASS doSomeGets(array) is 773
PASS counter is 351
PASS array[1] is 351
PASS doSomeGets(array) is 775
PASS counter is 352
PASS array[1] is 352
PASS doSomeGets(array) is 777
PASS counter is 353
PASS array[1] is 353
PASS doSomeGets(array) is 779
PASS counter is 354
PASS array[1] is 354
PASS doSomeGets(array) is 781
PASS counter is 355
PASS array[1] is 355
PASS doSomeGets(array) is 783
PASS counter is 356
PASS array[1] is 356
PASS doSomeGets(array) is 785
PASS counter is 357
PASS array[1] is 357
PASS doSomeGets(array) is 787
PASS counter is 358
PASS array[1] is 358
PASS doSomeGets(array) is 789
PASS counter is 359
PASS array[1] is 359
PASS doSomeGets(array) is 791
PASS counter is 360
PASS array[1] is 360
PASS doSomeGets(array) is 793
PASS counter is 361
PASS array[1] is 361
PASS doSomeGets(array) is 795
PASS counter is 362
PASS array[1] is 362
PASS doSomeGets(array) is 797
PASS counter is 363
PASS array[1] is 363
PASS doSomeGets(array) is 799
PASS counter is 364
PASS array[1] is 364
PASS doSomeGets(array) is 801
PASS counter is 365
PASS array[1] is 365
PASS doSomeGets(array) is 803
PASS counter is 366
PASS array[1] is 366
PASS doSomeGets(array) is 805
PASS counter is 367
PASS array[1] is 367
PASS doSomeGets(array) is 807
PASS counter is 368
PASS array[1] is 368
PASS doSomeGets(array) is 809
PASS counter is 369
PASS array[1] is 369
PASS doSomeGets(array) is 811
PASS counter is 370
PASS array[1] is 370
PASS doSomeGets(array) is 813
PASS counter is 371
PASS array[1] is 371
PASS doSomeGets(array) is 815
PASS counter is 372
PASS array[1] is 372
PASS doSomeGets(array) is 817
PASS counter is 373
PASS array[1] is 373
PASS doSomeGets(array) is 819
PASS counter is 374
PASS array[1] is 374
PASS doSomeGets(array) is 821
PASS counter is 375
PASS array[1] is 375
PASS doSomeGets(array) is 823
PASS counter is 376
PASS array[1] is 376
PASS doSomeGets(array) is 825
PASS counter is 377
PASS array[1] is 377
PASS doSomeGets(array) is 827
PASS counter is 378
PASS array[1] is 378
PASS doSomeGets(array) is 829
PASS counter is 379
PASS array[1] is 379
PASS doSomeGets(array) is 831
PASS counter is 380
PASS array[1] is 380
PASS doSomeGets(array) is 833
PASS counter is 381
PASS array[1] is 381
PASS doSomeGets(array) is 835
PASS counter is 382
PASS array[1] is 382
PASS doSomeGets(array) is 837
PASS counter is 383
PASS array[1] is 383
PASS doSomeGets(array) is 839
PASS counter is 384
PASS array[1] is 384
PASS doSomeGets(array) is 841
PASS counter is 385
PASS array[1] is 385
PASS doSomeGets(array) is 843
PASS counter is 386
PASS array[1] is 386
PASS doSomeGets(array) is 845
PASS counter is 387
PASS array[1] is 387
PASS doSomeGets(array) is 847
PASS counter is 388
PASS array[1] is 388
PASS doSomeGets(array) is 849
PASS counter is 389
PASS array[1] is 389
PASS doSomeGets(array) is 851
PASS counter is 390
PASS array[1] is 390
PASS doSomeGets(array) is 853
PASS counter is 391
PASS array[1] is 391
PASS doSomeGets(array) is 855
PASS counter is 392
PASS array[1] is 392
PASS doSomeGets(array) is 857
PASS counter is 393
PASS array[1] is 393
PASS doSomeGets(array) is 859
PASS counter is 394
PASS array[1] is 394
PASS doSomeGets(array) is 861
PASS counter is 395
PASS array[1] is 395
PASS doSomeGets(array) is 863
PASS counter is 396
PASS array[1] is 396
PASS doSomeGets(array) is 865
PASS counter is 397
PASS array[1] is 397
PASS doSomeGets(array) is 867
PASS counter is 398
PASS array[1] is 398
PASS doSomeGets(array) is 869
PASS counter is 399
PASS array[1] is 399
PASS doSomeGets(array) is 871
PASS counter is 400
PASS array[1] is 400
PASS doSomeGets(array) is 873
PASS counter is 401
PASS array[1] is 401
PASS doSomeGets(array) is 875
PASS counter is 402
PASS array[1] is 402
PASS doSomeGets(array) is 877
PASS counter is 403
PASS array[1] is 403
PASS doSomeGets(array) is 879
PASS counter is 404
PASS array[1] is 404
PASS doSomeGets(array) is 881
PASS counter is 405
PASS array[1] is 405
PASS doSomeGets(array) is 883
PASS counter is 406
PASS array[1] is 406
PASS doSomeGets(array) is 885
PASS counter is 407
PASS array[1] is 407
PASS doSomeGets(array) is 887
PASS counter is 408
PASS array[1] is 408
PASS doSomeGets(array) is 889
PASS counter is 409
PASS array[1] is 409
PASS doSomeGets(array) is 891
PASS counter is 410
PASS array[1] is 410
PASS doSomeGets(array) is 893
PASS counter is 411
PASS array[1] is 411
PASS doSomeGets(array) is 895
PASS counter is 412
PASS array[1] is 412
PASS doSomeGets(array) is 897
PASS counter is 413
PASS array[1] is 413
PASS doSomeGets(array) is 899
PASS counter is 414
PASS array[1] is 414
PASS doSomeGets(array) is 901
PASS counter is 415
PASS array[1] is 415
PASS doSomeGets(array) is 903
PASS counter is 416
PASS array[1] is 416
PASS doSomeGets(array) is 905
PASS counter is 417
PASS array[1] is 417
PASS doSomeGets(array) is 907
PASS counter is 418
PASS array[1] is 418
PASS doSomeGets(array) is 909
PASS counter is 419
PASS array[1] is 419
PASS doSomeGets(array) is 911
PASS counter is 420
PASS array[1] is 420
PASS doSomeGets(array) is 913
PASS counter is 421
PASS array[1] is 421
PASS doSomeGets(array) is 915
PASS counter is 422
PASS array[1] is 422
PASS doSomeGets(array) is 917
PASS counter is 423
PASS array[1] is 423
PASS doSomeGets(array) is 919
PASS counter is 424
PASS array[1] is 424
PASS doSomeGets(array) is 921
PASS counter is 425
PASS array[1] is 425
PASS doSomeGets(array) is 923
PASS counter is 426
PASS array[1] is 426
PASS doSomeGets(array) is 925
PASS counter is 427
PASS array[1] is 427
PASS doSomeGets(array) is 927
PASS counter is 428
PASS array[1] is 428
PASS doSomeGets(array) is 929
PASS counter is 429
PASS array[1] is 429
PASS doSomeGets(array) is 931
PASS counter is 430
PASS array[1] is 430
PASS doSomeGets(array) is 933
PASS counter is 431
PASS array[1] is 431
PASS doSomeGets(array) is 935
PASS counter is 432
PASS array[1] is 432
PASS doSomeGets(array) is 937
PASS counter is 433
PASS array[1] is 433
PASS doSomeGets(array) is 939
PASS counter is 434
PASS array[1] is 434
PASS doSomeGets(array) is 941
PASS counter is 435
PASS array[1] is 435
PASS doSomeGets(array) is 943
PASS counter is 436
PASS array[1] is 436
PASS doSomeGets(array) is 945
PASS counter is 437
PASS array[1] is 437
PASS doSomeGets(array) is 947
PASS counter is 438
PASS array[1] is 438
PASS doSomeGets(array) is 949
PASS counter is 439
PASS array[1] is 439
PASS doSomeGets(array) is 951
PASS counter is 440
PASS array[1] is 440
PASS doSomeGets(array) is 953
PASS counter is 441
PASS array[1] is 441
PASS doSomeGets(array) is 955
PASS counter is 442
PASS array[1] is 442
PASS doSomeGets(array) is 957
PASS counter is 443
PASS array[1] is 443
PASS doSomeGets(array) is 959
PASS counter is 444
PASS array[1] is 444
PASS doSomeGets(array) is 961
PASS counter is 445
PASS array[1] is 445
PASS doSomeGets(array) is 963
PASS counter is 446
PASS array[1] is 446
PASS doSomeGets(array) is 965
PASS counter is 447
PASS array[1] is 447
PASS doSomeGets(array) is 967
PASS counter is 448
PASS array[1] is 448
PASS doSomeGets(array) is 969
PASS counter is 449
PASS array[1] is 449
PASS doSomeGets(array) is 971
PASS counter is 450
PASS array[1] is 450
PASS doSomeGets(array) is 973
PASS counter is 451
PASS array[1] is 451
PASS doSomeGets(array) is 975
PASS counter is 452
PASS array[1] is 452
PASS doSomeGets(array) is 977
PASS counter is 453
PASS array[1] is 453
PASS doSomeGets(array) is 979
PASS counter is 454
PASS array[1] is 454
PASS doSomeGets(array) is 981
PASS counter is 455
PASS array[1] is 455
PASS doSomeGets(array) is 983
PASS counter is 456
PASS array[1] is 456
PASS doSomeGets(array) is 985
PASS counter is 457
PASS array[1] is 457
PASS doSomeGets(array) is 987
PASS counter is 458
PASS array[1] is 458
PASS doSomeGets(array) is 989
PASS counter is 459
PASS array[1] is 459
PASS doSomeGets(array) is 991
PASS counter is 460
PASS array[1] is 460
PASS doSomeGets(array) is 993
PASS counter is 461
PASS array[1] is 461
PASS doSomeGets(array) is 995
PASS counter is 462
PASS array[1] is 462
PASS doSomeGets(array) is 997
PASS counter is 463
PASS array[1] is 463
PASS doSomeGets(array) is 999
PASS counter is 464
PASS array[1] is 464
PASS doSomeGets(array) is 1001
PASS counter is 465
PASS array[1] is 465
PASS doSomeGets(array) is 1003
PASS counter is 466
PASS array[1] is 466
PASS doSomeGets(array) is 1005
PASS counter is 467
PASS array[1] is 467
PASS doSomeGets(array) is 1007
PASS counter is 468
PASS array[1] is 468
PASS doSomeGets(array) is 1009
PASS counter is 469
PASS array[1] is 469
PASS doSomeGets(array) is 1011
PASS counter is 470
PASS array[1] is 470
PASS doSomeGets(array) is 1013
PASS counter is 471
PASS array[1] is 471
PASS doSomeGets(array) is 1015
PASS counter is 472
PASS array[1] is 472
PASS doSomeGets(array) is 1017
PASS counter is 473
PASS array[1] is 473
PASS doSomeGets(array) is 1019
PASS counter is 474
PASS array[1] is 474
PASS doSomeGets(array) is 1021
PASS counter is 475
PASS array[1] is 475
PASS doSomeGets(array) is 1023
PASS counter is 476
PASS array[1] is 476
PASS doSomeGets(array) is 1025
PASS counter is 477
PASS array[1] is 477
PASS doSomeGets(array) is 1027
PASS counter is 478
PASS array[1] is 478
PASS doSomeGets(array) is 1029
PASS counter is 479
PASS array[1] is 479
PASS doSomeGets(array) is 1031
PASS counter is 480
PASS array[1] is 480
PASS doSomeGets(array) is 1033
PASS counter is 481
PASS array[1] is 481
PASS doSomeGets(array) is 1035
PASS counter is 482
PASS array[1] is 482
PASS doSomeGets(array) is 1037
PASS counter is 483
PASS array[1] is 483
PASS doSomeGets(array) is 1039
PASS counter is 484
PASS array[1] is 484
PASS doSomeGets(array) is 1041
PASS counter is 485
PASS array[1] is 485
PASS doSomeGets(array) is 1043
PASS counter is 486
PASS array[1] is 486
PASS doSomeGets(array) is 1045
PASS counter is 487
PASS array[1] is 487
PASS doSomeGets(array) is 1047
PASS counter is 488
PASS array[1] is 488
PASS doSomeGets(array) is 1049
PASS counter is 489
PASS array[1] is 489
PASS doSomeGets(array) is 1051
PASS counter is 490
PASS array[1] is 490
PASS doSomeGets(array) is 1053
PASS counter is 491
PASS array[1] is 491
PASS doSomeGets(array) is 1055
PASS counter is 492
PASS array[1] is 492
PASS doSomeGets(array) is 1057
PASS counter is 493
PASS array[1] is 493
PASS doSomeGets(array) is 1059
PASS counter is 494
PASS array[1] is 494
PASS doSomeGets(array) is 1061
PASS counter is 495
PASS array[1] is 495
PASS doSomeGets(array) is 1063
PASS counter is 496
PASS array[1] is 496
PASS doSomeGets(array) is 1065
PASS counter is 497
PASS array[1] is 497
PASS doSomeGets(array) is 1067
PASS counter is 498
PASS array[1] is 498
PASS doSomeGets(array) is 1069
PASS counter is 499
PASS array[1] is 499
PASS doSomeGets(array) is 1071
PASS counter is 500
PASS array[1] is 500
PASS doSomeGets(array) is 1073
PASS counter is 501
PASS array[1] is 501
PASS doSomeGets(array) is 1075
PASS counter is 502
PASS array[1] is 502
PASS doSomeGets(array) is 1077
PASS counter is 503
PASS array[1] is 503
PASS doSomeGets(array) is 1079
PASS counter is 504
PASS array[1] is 504
PASS doSomeGets(array) is 1081
PASS counter is 505
PASS array[1] is 505
PASS doSomeGets(array) is 1083
PASS counter is 506
PASS array[1] is 506
PASS doSomeGets(array) is 1085
PASS counter is 507
PASS array[1] is 507
PASS doSomeGets(array) is 1087
PASS counter is 508
PASS array[1] is 508
PASS doSomeGets(array) is 1089
PASS counter is 509
PASS array[1] is 509
PASS doSomeGets(array) is 1091
PASS counter is 510
PASS array[1] is 510
PASS doSomeGets(array) is 1093
PASS counter is 511
PASS array[1] is 511
PASS doSomeGets(array) is 1095
PASS counter is 512
PASS array[1] is 512
PASS doSomeGets(array) is 1097
PASS counter is 513
PASS array[1] is 513
PASS doSomeGets(array) is 1099
PASS counter is 514
PASS array[1] is 514
PASS doSomeGets(array) is 1101
PASS counter is 515
PASS array[1] is 515
PASS doSomeGets(array) is 1103
PASS counter is 516
PASS array[1] is 516
PASS doSomeGets(array) is 1105
PASS counter is 517
PASS array[1] is 517
PASS doSomeGets(array) is 1107
PASS counter is 518
PASS array[1] is 518
PASS doSomeGets(array) is 1109
PASS counter is 519
PASS array[1] is 519
PASS doSomeGets(array) is 1111
PASS counter is 520
PASS array[1] is 520
PASS doSomeGets(array) is 1113
PASS counter is 521
PASS array[1] is 521
PASS doSomeGets(array) is 1115
PASS counter is 522
PASS array[1] is 522
PASS doSomeGets(array) is 1117
PASS counter is 523
PASS array[1] is 523
PASS doSomeGets(array) is 1119
PASS counter is 524
PASS array[1] is 524
PASS doSomeGets(array) is 1121
PASS counter is 525
PASS array[1] is 525
PASS doSomeGets(array) is 1123
PASS counter is 526
PASS array[1] is 526
PASS doSomeGets(array) is 1125
PASS counter is 527
PASS array[1] is 527
PASS doSomeGets(array) is 1127
PASS counter is 528
PASS array[1] is 528
PASS doSomeGets(array) is 1129
PASS counter is 529
PASS array[1] is 529
PASS doSomeGets(array) is 1131
PASS counter is 530
PASS array[1] is 530
PASS doSomeGets(array) is 1133
PASS counter is 531
PASS array[1] is 531
PASS doSomeGets(array) is 1135
PASS counter is 532
PASS array[1] is 532
PASS doSomeGets(array) is 1137
PASS counter is 533
PASS array[1] is 533
PASS doSomeGets(array) is 1139
PASS counter is 534
PASS array[1] is 534
PASS doSomeGets(array) is 1141
PASS counter is 535
PASS array[1] is 535
PASS doSomeGets(array) is 1143
PASS counter is 536
PASS array[1] is 536
PASS doSomeGets(array) is 1145
PASS counter is 537
PASS array[1] is 537
PASS doSomeGets(array) is 1147
PASS counter is 538
PASS array[1] is 538
PASS doSomeGets(array) is 1149
PASS counter is 539
PASS array[1] is 539
PASS doSomeGets(array) is 1151
PASS counter is 540
PASS array[1] is 540
PASS doSomeGets(array) is 1153
PASS counter is 541
PASS array[1] is 541
PASS doSomeGets(array) is 1155
PASS counter is 542
PASS array[1] is 542
PASS doSomeGets(array) is 1157
PASS counter is 543
PASS array[1] is 543
PASS doSomeGets(array) is 1159
PASS counter is 544
PASS array[1] is 544
PASS doSomeGets(array) is 1161
PASS counter is 545
PASS array[1] is 545
PASS doSomeGets(array) is 1163
PASS counter is 546
PASS array[1] is 546
PASS doSomeGets(array) is 1165
PASS counter is 547
PASS array[1] is 547
PASS doSomeGets(array) is 1167
PASS counter is 548
PASS array[1] is 548
PASS doSomeGets(array) is 1169
PASS counter is 549
PASS array[1] is 549
PASS doSomeGets(array) is 1171
PASS counter is 550
PASS array[1] is 550
PASS doSomeGets(array) is 1173
PASS counter is 551
PASS array[1] is 551
PASS doSomeGets(array) is 1175
PASS counter is 552
PASS array[1] is 552
PASS doSomeGets(array) is 1177
PASS counter is 553
PASS array[1] is 553
PASS doSomeGets(array) is 1179
PASS counter is 554
PASS array[1] is 554
PASS doSomeGets(array) is 1181
PASS counter is 555
PASS array[1] is 555
PASS doSomeGets(array) is 1183
PASS counter is 556
PASS array[1] is 556
PASS doSomeGets(array) is 1185
PASS counter is 557
PASS array[1] is 557
PASS doSomeGets(array) is 1187
PASS counter is 558
PASS array[1] is 558
PASS doSomeGets(array) is 1189
PASS counter is 559
PASS array[1] is 559
PASS doSomeGets(array) is 1191
PASS counter is 560
PASS array[1] is 560
PASS doSomeGets(array) is 1193
PASS counter is 561
PASS array[1] is 561
PASS doSomeGets(array) is 1195
PASS counter is 562
PASS array[1] is 562
PASS doSomeGets(array) is 1197
PASS counter is 563
PASS array[1] is 563
PASS doSomeGets(array) is 1199
PASS counter is 564
PASS array[1] is 564
PASS doSomeGets(array) is 1201
PASS counter is 565
PASS array[1] is 565
PASS doSomeGets(array) is 1203
PASS counter is 566
PASS array[1] is 566
PASS doSomeGets(array) is 1205
PASS counter is 567
PASS array[1] is 567
PASS doSomeGets(array) is 1207
PASS counter is 568
PASS array[1] is 568
PASS doSomeGets(array) is 1209
PASS counter is 569
PASS array[1] is 569
PASS doSomeGets(array) is 1211
PASS counter is 570
PASS array[1] is 570
PASS doSomeGets(array) is 1213
PASS counter is 571
PASS array[1] is 571
PASS doSomeGets(array) is 1215
PASS counter is 572
PASS array[1] is 572
PASS doSomeGets(array) is 1217
PASS counter is 573
PASS array[1] is 573
PASS doSomeGets(array) is 1219
PASS counter is 574
PASS array[1] is 574
PASS doSomeGets(array) is 1221
PASS counter is 575
PASS array[1] is 575
PASS doSomeGets(array) is 1223
PASS counter is 576
PASS array[1] is 576
PASS doSomeGets(array) is 1225
PASS counter is 577
PASS array[1] is 577
PASS doSomeGets(array) is 1227
PASS counter is 578
PASS array[1] is 578
PASS doSomeGets(array) is 1229
PASS counter is 579
PASS array[1] is 579
PASS doSomeGets(array) is 1231
PASS counter is 580
PASS array[1] is 580
PASS doSomeGets(array) is 1233
PASS counter is 581
PASS array[1] is 581
PASS doSomeGets(array) is 1235
PASS counter is 582
PASS array[1] is 582
PASS doSomeGets(array) is 1237
PASS counter is 583
PASS array[1] is 583
PASS doSomeGets(array) is 1239
PASS counter is 584
PASS array[1] is 584
PASS doSomeGets(array) is 1241
PASS counter is 585
PASS array[1] is 585
PASS doSomeGets(array) is 1243
PASS counter is 586
PASS array[1] is 586
PASS doSomeGets(array) is 1245
PASS counter is 587
PASS array[1] is 587
PASS doSomeGets(array) is 1247
PASS counter is 588
PASS array[1] is 588
PASS doSomeGets(array) is 1249
PASS counter is 589
PASS array[1] is 589
PASS doSomeGets(array) is 1251
PASS counter is 590
PASS array[1] is 590
PASS doSomeGets(array) is 1253
PASS counter is 591
PASS array[1] is 591
PASS doSomeGets(array) is 1255
PASS counter is 592
PASS array[1] is 592
PASS doSomeGets(array) is 1257
PASS counter is 593
PASS array[1] is 593
PASS doSomeGets(array) is 1259
PASS counter is 594
PASS array[1] is 594
PASS doSomeGets(array) is 1261
PASS counter is 595
PASS array[1] is 595
PASS doSomeGets(array) is 1263
PASS counter is 596
PASS array[1] is 596
PASS doSomeGets(array) is 1265
PASS counter is 597
PASS array[1] is 597
PASS doSomeGets(array) is 1267
PASS counter is 598
PASS array[1] is 598
PASS doSomeGets(array) is 1269
PASS counter is 599
PASS array[1] is 599
PASS doSomeGets(array) is 1271
PASS counter is 600
PASS array[1] is 600
PASS doSomeGets(array) is 1273
PASS counter is 601
PASS array[1] is 601
PASS doSomeGets(array) is 1275
PASS counter is 602
PASS array[1] is 602
PASS doSomeGets(array) is 1277
PASS counter is 603
PASS array[1] is 603
PASS doSomeGets(array) is 1279
PASS counter is 604
PASS array[1] is 604
PASS doSomeGets(array) is 1281
PASS counter is 605
PASS array[1] is 605
PASS doSomeGets(array) is 1283
PASS counter is 606
PASS array[1] is 606
PASS doSomeGets(array) is 1285
PASS counter is 607
PASS array[1] is 607
PASS doSomeGets(array) is 1287
PASS counter is 608
PASS array[1] is 608
PASS doSomeGets(array) is 1289
PASS counter is 609
PASS array[1] is 609
PASS doSomeGets(array) is 1291
PASS counter is 610
PASS array[1] is 610
PASS doSomeGets(array) is 1293
PASS counter is 611
PASS array[1] is 611
PASS doSomeGets(array) is 1295
PASS counter is 612
PASS array[1] is 612
PASS doSomeGets(array) is 1297
PASS counter is 613
PASS array[1] is 613
PASS doSomeGets(array) is 1299
PASS counter is 614
PASS array[1] is 614
PASS doSomeGets(array) is 1301
PASS counter is 615
PASS array[1] is 615
PASS doSomeGets(array) is 1303
PASS counter is 616
PASS array[1] is 616
PASS doSomeGets(array) is 1305
PASS counter is 617
PASS array[1] is 617
PASS doSomeGets(array) is 1307
PASS counter is 618
PASS array[1] is 618
PASS doSomeGets(array) is 1309
PASS counter is 619
PASS array[1] is 619
PASS doSomeGets(array) is 1311
PASS counter is 620
PASS array[1] is 620
PASS doSomeGets(array) is 1313
PASS counter is 621
PASS array[1] is 621
PASS doSomeGets(array) is 1315
PASS counter is 622
PASS array[1] is 622
PASS doSomeGets(array) is 1317
PASS counter is 623
PASS array[1] is 623
PASS doSomeGets(array) is 1319
PASS counter is 624
PASS array[1] is 624
PASS doSomeGets(array) is 1321
PASS counter is 625
PASS array[1] is 625
PASS doSomeGets(array) is 1323
PASS counter is 626
PASS array[1] is 626
PASS doSomeGets(array) is 1325
PASS counter is 627
PASS array[1] is 627
PASS doSomeGets(array) is 1327
PASS counter is 628
PASS array[1] is 628
PASS doSomeGets(array) is 1329
PASS counter is 629
PASS array[1] is 629
PASS doSomeGets(array) is 1331
PASS counter is 630
PASS array[1] is 630
PASS doSomeGets(array) is 1333
PASS counter is 631
PASS array[1] is 631
PASS doSomeGets(array) is 1335
PASS counter is 632
PASS array[1] is 632
PASS doSomeGets(array) is 1337
PASS counter is 633
PASS array[1] is 633
PASS doSomeGets(array) is 1339
PASS counter is 634
PASS array[1] is 634
PASS doSomeGets(array) is 1341
PASS counter is 635
PASS array[1] is 635
PASS doSomeGets(array) is 1343
PASS counter is 636
PASS array[1] is 636
PASS doSomeGets(array) is 1345
PASS counter is 637
PASS array[1] is 637
PASS doSomeGets(array) is 1347
PASS counter is 638
PASS array[1] is 638
PASS doSomeGets(array) is 1349
PASS counter is 639
PASS array[1] is 639
PASS doSomeGets(array) is 1351
PASS counter is 640
PASS array[1] is 640
PASS doSomeGets(array) is 1353
PASS counter is 641
PASS array[1] is 641
PASS doSomeGets(array) is 1355
PASS counter is 642
PASS array[1] is 642
PASS doSomeGets(array) is 1357
PASS counter is 643
PASS array[1] is 643
PASS doSomeGets(array) is 1359
PASS counter is 644
PASS array[1] is 644
PASS doSomeGets(array) is 1361
PASS counter is 645
PASS array[1] is 645
PASS doSomeGets(array) is 1363
PASS counter is 646
PASS array[1] is 646
PASS doSomeGets(array) is 1365
PASS counter is 647
PASS array[1] is 647
PASS doSomeGets(array) is 1367
PASS counter is 648
PASS array[1] is 648
PASS doSomeGets(array) is 1369
PASS counter is 649
PASS array[1] is 649
PASS doSomeGets(array) is 1371
PASS counter is 650
PASS array[1] is 650
PASS doSomeGets(array) is 1373
PASS counter is 651
PASS array[1] is 651
PASS doSomeGets(array) is 1375
PASS counter is 652
PASS array[1] is 652
PASS doSomeGets(array) is 1377
PASS counter is 653
PASS array[1] is 653
PASS doSomeGets(array) is 1379
PASS counter is 654
PASS array[1] is 654
PASS doSomeGets(array) is 1381
PASS counter is 655
PASS array[1] is 655
PASS doSomeGets(array) is 1383
PASS counter is 656
PASS array[1] is 656
PASS doSomeGets(array) is 1385
PASS counter is 657
PASS array[1] is 657
PASS doSomeGets(array) is 1387
PASS counter is 658
PASS array[1] is 658
PASS doSomeGets(array) is 1389
PASS counter is 659
PASS array[1] is 659
PASS doSomeGets(array) is 1391
PASS counter is 660
PASS array[1] is 660
PASS doSomeGets(array) is 1393
PASS counter is 661
PASS array[1] is 661
PASS doSomeGets(array) is 1395
PASS counter is 662
PASS array[1] is 662
PASS doSomeGets(array) is 1397
PASS counter is 663
PASS array[1] is 663
PASS doSomeGets(array) is 1399
PASS counter is 664
PASS array[1] is 664
PASS doSomeGets(array) is 1401
PASS counter is 665
PASS array[1] is 665
PASS doSomeGets(array) is 1403
PASS counter is 666
PASS array[1] is 666
PASS doSomeGets(array) is 1405
PASS counter is 667
PASS array[1] is 667
PASS doSomeGets(array) is 1407
PASS counter is 668
PASS array[1] is 668
PASS doSomeGets(array) is 1409
PASS counter is 669
PASS array[1] is 669
PASS doSomeGets(array) is 1411
PASS counter is 670
PASS array[1] is 670
PASS doSomeGets(array) is 1413
PASS counter is 671
PASS array[1] is 671
PASS doSomeGets(array) is 1415
PASS counter is 672
PASS array[1] is 672
PASS doSomeGets(array) is 1417
PASS counter is 673
PASS array[1] is 673
PASS doSomeGets(array) is 1419
PASS counter is 674
PASS array[1] is 674
PASS doSomeGets(array) is 1421
PASS counter is 675
PASS array[1] is 675
PASS doSomeGets(array) is 1423
PASS counter is 676
PASS array[1] is 676
PASS doSomeGets(array) is 1425
PASS counter is 677
PASS array[1] is 677
PASS doSomeGets(array) is 1427
PASS counter is 678
PASS array[1] is 678
PASS doSomeGets(array) is 1429
PASS counter is 679
PASS array[1] is 679
PASS doSomeGets(array) is 1431
PASS counter is 680
PASS array[1] is 680
PASS doSomeGets(array) is 1433
PASS counter is 681
PASS array[1] is 681
PASS doSomeGets(array) is 1435
PASS counter is 682
PASS array[1] is 682
PASS doSomeGets(array) is 1437
PASS counter is 683
PASS array[1] is 683
PASS doSomeGets(array) is 1439
PASS counter is 684
PASS array[1] is 684
PASS doSomeGets(array) is 1441
PASS counter is 685
PASS array[1] is 685
PASS doSomeGets(array) is 1443
PASS counter is 686
PASS array[1] is 686
PASS doSomeGets(array) is 1445
PASS counter is 687
PASS array[1] is 687
PASS doSomeGets(array) is 1447
PASS counter is 688
PASS array[1] is 688
PASS doSomeGets(array) is 1449
PASS counter is 689
PASS array[1] is 689
PASS doSomeGets(array) is 1451
PASS counter is 690
PASS array[1] is 690
PASS doSomeGets(array) is 1453
PASS counter is 691
PASS array[1] is 691
PASS doSomeGets(array) is 1455
PASS counter is 692
PASS array[1] is 692
PASS doSomeGets(array) is 1457
PASS counter is 693
PASS array[1] is 693
PASS doSomeGets(array) is 1459
PASS counter is 694
PASS array[1] is 694
PASS doSomeGets(array) is 1461
PASS counter is 695
PASS array[1] is 695
PASS doSomeGets(array) is 1463
PASS counter is 696
PASS array[1] is 696
PASS doSomeGets(array) is 1465
PASS counter is 697
PASS array[1] is 697
PASS doSomeGets(array) is 1467
PASS counter is 698
PASS array[1] is 698
PASS doSomeGets(array) is 1469
PASS counter is 699
PASS array[1] is 699
PASS doSomeGets(array) is 1471
PASS counter is 700
PASS array[1] is 700
PASS doSomeGets(array) is 1473
PASS counter is 701
PASS array[1] is 701
PASS doSomeGets(array) is 1475
PASS counter is 702
PASS array[1] is 702
PASS doSomeGets(array) is 1477
PASS counter is 703
PASS array[1] is 703
PASS doSomeGets(array) is 1479
PASS counter is 704
PASS array[1] is 704
PASS doSomeGets(array) is 1481
PASS counter is 705
PASS array[1] is 705
PASS doSomeGets(array) is 1483
PASS counter is 706
PASS array[1] is 706
PASS doSomeGets(array) is 1485
PASS counter is 707
PASS array[1] is 707
PASS doSomeGets(array) is 1487
PASS counter is 708
PASS array[1] is 708
PASS doSomeGets(array) is 1489
PASS counter is 709
PASS array[1] is 709
PASS doSomeGets(array) is 1491
PASS counter is 710
PASS array[1] is 710
PASS doSomeGets(array) is 1493
PASS counter is 711
PASS array[1] is 711
PASS doSomeGets(array) is 1495
PASS counter is 712
PASS array[1] is 712
PASS doSomeGets(array) is 1497
PASS counter is 713
PASS array[1] is 713
PASS doSomeGets(array) is 1499
PASS counter is 714
PASS array[1] is 714
PASS doSomeGets(array) is 1501
PASS counter is 715
PASS array[1] is 715
PASS doSomeGets(array) is 1503
PASS counter is 716
PASS array[1] is 716
PASS doSomeGets(array) is 1505
PASS counter is 717
PASS array[1] is 717
PASS doSomeGets(array) is 1507
PASS counter is 718
PASS array[1] is 718
PASS doSomeGets(array) is 1509
PASS counter is 719
PASS array[1] is 719
PASS doSomeGets(array) is 1511
PASS counter is 720
PASS array[1] is 720
PASS doSomeGets(array) is 1513
PASS counter is 721
PASS array[1] is 721
PASS doSomeGets(array) is 1515
PASS counter is 722
PASS array[1] is 722
PASS doSomeGets(array) is 1517
PASS counter is 723
PASS array[1] is 723
PASS doSomeGets(array) is 1519
PASS counter is 724
PASS array[1] is 724
PASS doSomeGets(array) is 1521
PASS counter is 725
PASS array[1] is 725
PASS doSomeGets(array) is 1523
PASS counter is 726
PASS array[1] is 726
PASS doSomeGets(array) is 1525
PASS counter is 727
PASS array[1] is 727
PASS doSomeGets(array) is 1527
PASS counter is 728
PASS array[1] is 728
PASS doSomeGets(array) is 1529
PASS counter is 729
PASS array[1] is 729
PASS doSomeGets(array) is 1531
PASS counter is 730
PASS array[1] is 730
PASS doSomeGets(array) is 1533
PASS counter is 731
PASS array[1] is 731
PASS doSomeGets(array) is 1535
PASS counter is 732
PASS array[1] is 732
PASS doSomeGets(array) is 1537
PASS counter is 733
PASS array[1] is 733
PASS doSomeGets(array) is 1539
PASS counter is 734
PASS array[1] is 734
PASS doSomeGets(array) is 1541
PASS counter is 735
PASS array[1] is 735
PASS doSomeGets(array) is 1543
PASS counter is 736
PASS array[1] is 736
PASS doSomeGets(array) is 1545
PASS counter is 737
PASS array[1] is 737
PASS doSomeGets(array) is 1547
PASS counter is 738
PASS array[1] is 738
PASS doSomeGets(array) is 1549
PASS counter is 739
PASS array[1] is 739
PASS doSomeGets(array) is 1551
PASS counter is 740
PASS array[1] is 740
PASS doSomeGets(array) is 1553
PASS counter is 741
PASS array[1] is 741
PASS doSomeGets(array) is 1555
PASS counter is 742
PASS array[1] is 742
PASS doSomeGets(array) is 1557
PASS counter is 743
PASS array[1] is 743
PASS doSomeGets(array) is 1559
PASS counter is 744
PASS array[1] is 744
PASS doSomeGets(array) is 1561
PASS counter is 745
PASS array[1] is 745
PASS doSomeGets(array) is 1563
PASS counter is 746
PASS array[1] is 746
PASS doSomeGets(array) is 1565
PASS counter is 747
PASS array[1] is 747
PASS doSomeGets(array) is 1567
PASS counter is 748
PASS array[1] is 748
PASS doSomeGets(array) is 1569
PASS counter is 749
PASS array[1] is 749
PASS doSomeGets(array) is 1571
PASS counter is 750
PASS array[1] is 750
PASS doSomeGets(array) is 1573
PASS counter is 751
PASS array[1] is 751
PASS doSomeGets(array) is 1575
PASS counter is 752
PASS array[1] is 752
PASS doSomeGets(array) is 1577
PASS counter is 753
PASS array[1] is 753
PASS doSomeGets(array) is 1579
PASS counter is 754
PASS array[1] is 754
PASS doSomeGets(array) is 1581
PASS counter is 755
PASS array[1] is 755
PASS doSomeGets(array) is 1583
PASS counter is 756
PASS array[1] is 756
PASS doSomeGets(array) is 1585
PASS counter is 757
PASS array[1] is 757
PASS doSomeGets(array) is 1587
PASS counter is 758
PASS array[1] is 758
PASS doSomeGets(array) is 1589
PASS counter is 759
PASS array[1] is 759
PASS doSomeGets(array) is 1591
PASS counter is 760
PASS array[1] is 760
PASS doSomeGets(array) is 1593
PASS counter is 761
PASS array[1] is 761
PASS doSomeGets(array) is 1595
PASS counter is 762
PASS array[1] is 762
PASS doSomeGets(array) is 1597
PASS counter is 763
PASS array[1] is 763
PASS doSomeGets(array) is 1599
PASS counter is 764
PASS array[1] is 764
PASS doSomeGets(array) is 1601
PASS counter is 765
PASS array[1] is 765
PASS doSomeGets(array) is 1603
PASS counter is 766
PASS array[1] is 766
PASS doSomeGets(array) is 1605
PASS counter is 767
PASS array[1] is 767
PASS doSomeGets(array) is 1607
PASS counter is 768
PASS array[1] is 768
PASS doSomeGets(array) is 1609
PASS counter is 769
PASS array[1] is 769
PASS doSomeGets(array) is 1611
PASS counter is 770
PASS array[1] is 770
PASS doSomeGets(array) is 1613
PASS counter is 771
PASS array[1] is 771
PASS doSomeGets(array) is 1615
PASS counter is 772
PASS array[1] is 772
PASS doSomeGets(array) is 1617
PASS counter is 773
PASS array[1] is 773
PASS doSomeGets(array) is 1619
PASS counter is 774
PASS array[1] is 774
PASS doSomeGets(array) is 1621
PASS counter is 775
PASS array[1] is 775
PASS doSomeGets(array) is 1623
PASS counter is 776
PASS array[1] is 776
PASS doSomeGets(array) is 1625
PASS counter is 777
PASS array[1] is 777
PASS doSomeGets(array) is 1627
PASS counter is 778
PASS array[1] is 778
PASS doSomeGets(array) is 1629
PASS counter is 779
PASS array[1] is 779
PASS doSomeGets(array) is 1631
PASS counter is 780
PASS array[1] is 780
PASS doSomeGets(array) is 1633
PASS counter is 781
PASS array[1] is 781
PASS doSomeGets(array) is 1635
PASS counter is 782
PASS array[1] is 782
PASS doSomeGets(array) is 1637
PASS counter is 783
PASS array[1] is 783
PASS doSomeGets(array) is 1639
PASS counter is 784
PASS array[1] is 784
PASS doSomeGets(array) is 1641
PASS counter is 785
PASS array[1] is 785
PASS doSomeGets(array) is 1643
PASS counter is 786
PASS array[1] is 786
PASS doSomeGets(array) is 1645
PASS counter is 787
PASS array[1] is 787
PASS doSomeGets(array) is 1647
PASS counter is 788
PASS array[1] is 788
PASS doSomeGets(array) is 1649
PASS counter is 789
PASS array[1] is 789
PASS doSomeGets(array) is 1651
PASS counter is 790
PASS array[1] is 790
PASS doSomeGets(array) is 1653
PASS counter is 791
PASS array[1] is 791
PASS doSomeGets(array) is 1655
PASS counter is 792
PASS array[1] is 792
PASS doSomeGets(array) is 1657
PASS counter is 793
PASS array[1] is 793
PASS doSomeGets(array) is 1659
PASS counter is 794
PASS array[1] is 794
PASS doSomeGets(array) is 1661
PASS counter is 795
PASS array[1] is 795
PASS doSomeGets(array) is 1663
PASS counter is 796
PASS array[1] is 796
PASS doSomeGets(array) is 1665
PASS counter is 797
PASS array[1] is 797
PASS doSomeGets(array) is 1667
PASS counter is 798
PASS array[1] is 798
PASS doSomeGets(array) is 1669
PASS counter is 799
PASS array[1] is 799
PASS doSomeGets(array) is 1671
PASS counter is 800
PASS array[1] is 800
PASS doSomeGets(array) is 1673
PASS counter is 801
PASS array[1] is 801
PASS doSomeGets(array) is 1675
PASS counter is 802
PASS array[1] is 802
PASS doSomeGets(array) is 1677
PASS counter is 803
PASS array[1] is 803
PASS doSomeGets(array) is 1679
PASS counter is 804
PASS array[1] is 804
PASS doSomeGets(array) is 1681
PASS counter is 805
PASS array[1] is 805
PASS doSomeGets(array) is 1683
PASS counter is 806
PASS array[1] is 806
PASS doSomeGets(array) is 1685
PASS counter is 807
PASS array[1] is 807
PASS doSomeGets(array) is 1687
PASS counter is 808
PASS array[1] is 808
PASS doSomeGets(array) is 1689
PASS counter is 809
PASS array[1] is 809
PASS doSomeGets(array) is 1691
PASS counter is 810
PASS array[1] is 810
PASS doSomeGets(array) is 1693
PASS counter is 811
PASS array[1] is 811
PASS doSomeGets(array) is 1695
PASS counter is 812
PASS array[1] is 812
PASS doSomeGets(array) is 1697
PASS counter is 813
PASS array[1] is 813
PASS doSomeGets(array) is 1699
PASS counter is 814
PASS array[1] is 814
PASS doSomeGets(array) is 1701
PASS counter is 815
PASS array[1] is 815
PASS doSomeGets(array) is 1703
PASS counter is 816
PASS array[1] is 816
PASS doSomeGets(array) is 1705
PASS counter is 817
PASS array[1] is 817
PASS doSomeGets(array) is 1707
PASS counter is 818
PASS array[1] is 818
PASS doSomeGets(array) is 1709
PASS counter is 819
PASS array[1] is 819
PASS doSomeGets(array) is 1711
PASS counter is 820
PASS array[1] is 820
PASS doSomeGets(array) is 1713
PASS counter is 821
PASS array[1] is 821
PASS doSomeGets(array) is 1715
PASS counter is 822
PASS array[1] is 822
PASS doSomeGets(array) is 1717
PASS counter is 823
PASS array[1] is 823
PASS doSomeGets(array) is 1719
PASS counter is 824
PASS array[1] is 824
PASS doSomeGets(array) is 1721
PASS counter is 825
PASS array[1] is 825
PASS doSomeGets(array) is 1723
PASS counter is 826
PASS array[1] is 826
PASS doSomeGets(array) is 1725
PASS counter is 827
PASS array[1] is 827
PASS doSomeGets(array) is 1727
PASS counter is 828
PASS array[1] is 828
PASS doSomeGets(array) is 1729
PASS counter is 829
PASS array[1] is 829
PASS doSomeGets(array) is 1731
PASS counter is 830
PASS array[1] is 830
PASS doSomeGets(array) is 1733
PASS counter is 831
PASS array[1] is 831
PASS doSomeGets(array) is 1735
PASS counter is 832
PASS array[1] is 832
PASS doSomeGets(array) is 1737
PASS counter is 833
PASS array[1] is 833
PASS doSomeGets(array) is 1739
PASS counter is 834
PASS array[1] is 834
PASS doSomeGets(array) is 1741
PASS counter is 835
PASS array[1] is 835
PASS doSomeGets(array) is 1743
PASS counter is 836
PASS array[1] is 836
PASS doSomeGets(array) is 1745
PASS counter is 837
PASS array[1] is 837
PASS doSomeGets(array) is 1747
PASS counter is 838
PASS array[1] is 838
PASS doSomeGets(array) is 1749
PASS counter is 839
PASS array[1] is 839
PASS doSomeGets(array) is 1751
PASS counter is 840
PASS array[1] is 840
PASS doSomeGets(array) is 1753
PASS counter is 841
PASS array[1] is 841
PASS doSomeGets(array) is 1755
PASS counter is 842
PASS array[1] is 842
PASS doSomeGets(array) is 1757
PASS counter is 843
PASS array[1] is 843
PASS doSomeGets(array) is 1759
PASS counter is 844
PASS array[1] is 844
PASS doSomeGets(array) is 1761
PASS counter is 845
PASS array[1] is 845
PASS doSomeGets(array) is 1763
PASS counter is 846
PASS array[1] is 846
PASS doSomeGets(array) is 1765
PASS counter is 847
PASS array[1] is 847
PASS doSomeGets(array) is 1767
PASS counter is 848
PASS array[1] is 848
PASS doSomeGets(array) is 1769
PASS counter is 849
PASS array[1] is 849
PASS doSomeGets(array) is 1771
PASS counter is 850
PASS array[1] is 850
PASS doSomeGets(array) is 1773
PASS counter is 851
PASS array[1] is 851
PASS doSomeGets(array) is 1775
PASS counter is 852
PASS array[1] is 852
PASS doSomeGets(array) is 1777
PASS counter is 853
PASS array[1] is 853
PASS doSomeGets(array) is 1779
PASS counter is 854
PASS array[1] is 854
PASS doSomeGets(array) is 1781
PASS counter is 855
PASS array[1] is 855
PASS doSomeGets(array) is 1783
PASS counter is 856
PASS array[1] is 856
PASS doSomeGets(array) is 1785
PASS counter is 857
PASS array[1] is 857
PASS doSomeGets(array) is 1787
PASS counter is 858
PASS array[1] is 858
PASS doSomeGets(array) is 1789
PASS counter is 859
PASS array[1] is 859
PASS doSomeGets(array) is 1791
PASS counter is 860
PASS array[1] is 860
PASS doSomeGets(array) is 1793
PASS counter is 861
PASS array[1] is 861
PASS doSomeGets(array) is 1795
PASS counter is 862
PASS array[1] is 862
PASS doSomeGets(array) is 1797
PASS counter is 863
PASS array[1] is 863
PASS doSomeGets(array) is 1799
PASS counter is 864
PASS array[1] is 864
PASS doSomeGets(array) is 1801
PASS counter is 865
PASS array[1] is 865
PASS doSomeGets(array) is 1803
PASS counter is 866
PASS array[1] is 866
PASS doSomeGets(array) is 1805
PASS counter is 867
PASS array[1] is 867
PASS doSomeGets(array) is 1807
PASS counter is 868
PASS array[1] is 868
PASS doSomeGets(array) is 1809
PASS counter is 869
PASS array[1] is 869
PASS doSomeGets(array) is 1811
PASS counter is 870
PASS array[1] is 870
PASS doSomeGets(array) is 1813
PASS counter is 871
PASS array[1] is 871
PASS doSomeGets(array) is 1815
PASS counter is 872
PASS array[1] is 872
PASS doSomeGets(array) is 1817
PASS counter is 873
PASS array[1] is 873
PASS doSomeGets(array) is 1819
PASS counter is 874
PASS array[1] is 874
PASS doSomeGets(array) is 1821
PASS counter is 875
PASS array[1] is 875
PASS doSomeGets(array) is 1823
PASS counter is 876
PASS array[1] is 876
PASS doSomeGets(array) is 1825
PASS counter is 877
PASS array[1] is 877
PASS doSomeGets(array) is 1827
PASS counter is 878
PASS array[1] is 878
PASS doSomeGets(array) is 1829
PASS counter is 879
PASS array[1] is 879
PASS doSomeGets(array) is 1831
PASS counter is 880
PASS array[1] is 880
PASS doSomeGets(array) is 1833
PASS counter is 881
PASS array[1] is 881
PASS doSomeGets(array) is 1835
PASS counter is 882
PASS array[1] is 882
PASS doSomeGets(array) is 1837
PASS counter is 883
PASS array[1] is 883
PASS doSomeGets(array) is 1839
PASS counter is 884
PASS array[1] is 884
PASS doSomeGets(array) is 1841
PASS counter is 885
PASS array[1] is 885
PASS doSomeGets(array) is 1843
PASS counter is 886
PASS array[1] is 886
PASS doSomeGets(array) is 1845
PASS counter is 887
PASS array[1] is 887
PASS doSomeGets(array) is 1847
PASS counter is 888
PASS array[1] is 888
PASS doSomeGets(array) is 1849
PASS counter is 889
PASS array[1] is 889
PASS doSomeGets(array) is 1851
PASS counter is 890
PASS array[1] is 890
PASS doSomeGets(array) is 1853
PASS counter is 891
PASS array[1] is 891
PASS doSomeGets(array) is 1855
PASS counter is 892
PASS array[1] is 892
PASS doSomeGets(array) is 1857
PASS counter is 893
PASS array[1] is 893
PASS doSomeGets(array) is 1859
PASS counter is 894
PASS array[1] is 894
PASS doSomeGets(array) is 1861
PASS counter is 895
PASS array[1] is 895
PASS doSomeGets(array) is 1863
PASS counter is 896
PASS array[1] is 896
PASS doSomeGets(array) is 1865
PASS counter is 897
PASS array[1] is 897
PASS doSomeGets(array) is 1867
PASS counter is 898
PASS array[1] is 898
PASS doSomeGets(array) is 1869
PASS counter is 899
PASS array[1] is 899
PASS doSomeGets(array) is 1871
PASS counter is 900
PASS array[1] is 900
PASS doSomeGets(array) is 1873
PASS counter is 901
PASS array[1] is 901
PASS doSomeGets(array) is 1875
PASS counter is 902
PASS array[1] is 902
PASS doSomeGets(array) is 1877
PASS counter is 903
PASS array[1] is 903
PASS doSomeGets(array) is 1879
PASS counter is 904
PASS array[1] is 904
PASS doSomeGets(array) is 1881
PASS counter is 905
PASS array[1] is 905
PASS doSomeGets(array) is 1883
PASS counter is 906
PASS array[1] is 906
PASS doSomeGets(array) is 1885
PASS counter is 907
PASS array[1] is 907
PASS doSomeGets(array) is 1887
PASS counter is 908
PASS array[1] is 908
PASS doSomeGets(array) is 1889
PASS counter is 909
PASS array[1] is 909
PASS doSomeGets(array) is 1891
PASS counter is 910
PASS array[1] is 910
PASS doSomeGets(array) is 1893
PASS counter is 911
PASS array[1] is 911
PASS doSomeGets(array) is 1895
PASS counter is 912
PASS array[1] is 912
PASS doSomeGets(array) is 1897
PASS counter is 913
PASS array[1] is 913
PASS doSomeGets(array) is 1899
PASS counter is 914
PASS array[1] is 914
PASS doSomeGets(array) is 1901
PASS counter is 915
PASS array[1] is 915
PASS doSomeGets(array) is 1903
PASS counter is 916
PASS array[1] is 916
PASS doSomeGets(array) is 1905
PASS counter is 917
PASS array[1] is 917
PASS doSomeGets(array) is 1907
PASS counter is 918
PASS array[1] is 918
PASS doSomeGets(array) is 1909
PASS counter is 919
PASS array[1] is 919
PASS doSomeGets(array) is 1911
PASS counter is 920
PASS array[1] is 920
PASS doSomeGets(array) is 1913
PASS counter is 921
PASS array[1] is 921
PASS doSomeGets(array) is 1915
PASS counter is 922
PASS array[1] is 922
PASS doSomeGets(array) is 1917
PASS counter is 923
PASS array[1] is 923
PASS doSomeGets(array) is 1919
PASS counter is 924
PASS array[1] is 924
PASS doSomeGets(array) is 1921
PASS counter is 925
PASS array[1] is 925
PASS doSomeGets(array) is 1923
PASS counter is 926
PASS array[1] is 926
PASS doSomeGets(array) is 1925
PASS counter is 927
PASS array[1] is 927
PASS doSomeGets(array) is 1927
PASS counter is 928
PASS array[1] is 928
PASS doSomeGets(array) is 1929
PASS counter is 929
PASS array[1] is 929
PASS doSomeGets(array) is 1931
PASS counter is 930
PASS array[1] is 930
PASS doSomeGets(array) is 1933
PASS counter is 931
PASS array[1] is 931
PASS doSomeGets(array) is 1935
PASS counter is 932
PASS array[1] is 932
PASS doSomeGets(array) is 1937
PASS counter is 933
PASS array[1] is 933
PASS doSomeGets(array) is 1939
PASS counter is 934
PASS array[1] is 934
PASS doSomeGets(array) is 1941
PASS counter is 935
PASS array[1] is 935
PASS doSomeGets(array) is 1943
PASS counter is 936
PASS array[1] is 936
PASS doSomeGets(array) is 1945
PASS counter is 937
PASS array[1] is 937
PASS doSomeGets(array) is 1947
PASS counter is 938
PASS array[1] is 938
PASS doSomeGets(array) is 1949
PASS counter is 939
PASS array[1] is 939
PASS doSomeGets(array) is 1951
PASS counter is 940
PASS array[1] is 940
PASS doSomeGets(array) is 1953
PASS counter is 941
PASS array[1] is 941
PASS doSomeGets(array) is 1955
PASS counter is 942
PASS array[1] is 942
PASS doSomeGets(array) is 1957
PASS counter is 943
PASS array[1] is 943
PASS doSomeGets(array) is 1959
PASS counter is 944
PASS array[1] is 944
PASS doSomeGets(array) is 1961
PASS counter is 945
PASS array[1] is 945
PASS doSomeGets(array) is 1963
PASS counter is 946
PASS array[1] is 946
PASS doSomeGets(array) is 1965
PASS counter is 947
PASS array[1] is 947
PASS doSomeGets(array) is 1967
PASS counter is 948
PASS array[1] is 948
PASS doSomeGets(array) is 1969
PASS counter is 949
PASS array[1] is 949
PASS doSomeGets(array) is 1971
PASS counter is 950
PASS array[1] is 950
PASS doSomeGets(array) is 1973
PASS counter is 951
PASS array[1] is 951
PASS doSomeGets(array) is 1975
PASS counter is 952
PASS array[1] is 952
PASS doSomeGets(array) is 1977
PASS counter is 953
PASS array[1] is 953
PASS doSomeGets(array) is 1979
PASS counter is 954
PASS array[1] is 954
PASS doSomeGets(array) is 1981
PASS counter is 955
PASS array[1] is 955
PASS doSomeGets(array) is 1983
PASS counter is 956
PASS array[1] is 956
PASS doSomeGets(array) is 1985
PASS counter is 957
PASS array[1] is 957
PASS doSomeGets(array) is 1987
PASS counter is 958
PASS array[1] is 958
PASS doSomeGets(array) is 1989
PASS counter is 959
PASS array[1] is 959
PASS doSomeGets(array) is 1991
PASS counter is 960
PASS array[1] is 960
PASS doSomeGets(array) is 1993
PASS counter is 961
PASS array[1] is 961
PASS doSomeGets(array) is 1995
PASS counter is 962
PASS array[1] is 962
PASS doSomeGets(array) is 1997
PASS counter is 963
PASS array[1] is 963
PASS doSomeGets(array) is 1999
PASS counter is 964
PASS array[1] is 964
PASS doSomeGets(array) is 2001
PASS counter is 965
PASS array[1] is 965
PASS doSomeGets(array) is 2003
PASS counter is 966
PASS array[1] is 966
PASS doSomeGets(array) is 2005
PASS counter is 967
PASS array[1] is 967
PASS doSomeGets(array) is 2007
PASS counter is 968
PASS array[1] is 968
PASS doSomeGets(array) is 2009
PASS counter is 969
PASS array[1] is 969
PASS doSomeGets(array) is 2011
PASS counter is 970
PASS array[1] is 970
PASS doSomeGets(array) is 2013
PASS counter is 971
PASS array[1] is 971
PASS doSomeGets(array) is 2015
PASS counter is 972
PASS array[1] is 972
PASS doSomeGets(array) is 2017
PASS counter is 973
PASS array[1] is 973
PASS doSomeGets(array) is 2019
PASS counter is 974
PASS array[1] is 974
PASS doSomeGets(array) is 2021
PASS counter is 975
PASS array[1] is 975
PASS doSomeGets(array) is 2023
PASS counter is 976
PASS array[1] is 976
PASS doSomeGets(array) is 2025
PASS counter is 977
PASS array[1] is 977
PASS doSomeGets(array) is 2027
PASS counter is 978
PASS array[1] is 978
PASS doSomeGets(array) is 2029
PASS counter is 979
PASS array[1] is 979
PASS doSomeGets(array) is 2031
PASS counter is 980
PASS array[1] is 980
PASS doSomeGets(array) is 2033
PASS counter is 981
PASS array[1] is 981
PASS doSomeGets(array) is 2035
PASS counter is 982
PASS array[1] is 982
PASS doSomeGets(array) is 2037
PASS counter is 983
PASS array[1] is 983
PASS doSomeGets(array) is 2039
PASS counter is 984
PASS array[1] is 984
PASS doSomeGets(array) is 2041
PASS counter is 985
PASS array[1] is 985
PASS doSomeGets(array) is 2043
PASS counter is 986
PASS array[1] is 986
PASS doSomeGets(array) is 2045
PASS counter is 987
PASS array[1] is 987
PASS doSomeGets(array) is 2047
PASS counter is 988
PASS array[1] is 988
PASS doSomeGets(array) is 2049
PASS counter is 989
PASS array[1] is 989
PASS doSomeGets(array) is 2051
PASS counter is 990
PASS array[1] is 990
PASS doSomeGets(array) is 2053
PASS counter is 991
PASS array[1] is 991
PASS doSomeGets(array) is 2055
PASS counter is 992
PASS array[1] is 992
PASS doSomeGets(array) is 2057
PASS counter is 993
PASS array[1] is 993
PASS doSomeGets(array) is 2059
PASS counter is 994
PASS array[1] is 994
PASS doSomeGets(array) is 2061
PASS counter is 995
PASS array[1] is 995
PASS doSomeGets(array) is 2063
PASS counter is 996
PASS array[1] is 996
PASS doSomeGets(array) is 2065
PASS counter is 997
PASS array[1] is 997
PASS doSomeGets(array) is 2067
PASS counter is 998
PASS array[1] is 998
PASS doSomeGets(array) is 2069
PASS counter is 999
PASS array[1] is 999
PASS doSomeGets(array) is 2071
PASS counter is 1000
PASS array[1] is 1000
PASS doSomeGets(array) is 2073
PASS counter is 1001
PASS array[1] is 1001
PASS doSomeGets(array) is 2075
PASS counter is 1002
PASS array[1] is 1002
PASS successfullyParsed is true

TEST COMPLETE
